Equipment System
No icons and meshes right now for the separate items, but there is a data table which contains all the items. Then you create a BP_Weapon, where you can set which item it is and place it in the world. Also, in the inventory, you can select and equip items based on their category.